backend/requirements.txt
# ============================================
# BACKEND - REQUIREMENTS.TXT
# ============================================

# Framework web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de datos
supabase==2.3.0
psycopg2-binary==2.9.9

# Seguridad y autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# HTTP requests
requests==2.31.0
httpx==0.26.0

# AI y LangChain
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# Utilidades
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS y middleware
python-cors==1.0.0

# Logging y monitoring
python-json-logger==2.0.7
